Fuse into irc.py again
